modify credentialAtomicQueryMTP
